Adsorbed sensor_array_mof_adsorption.py into write_simulations.py.
